home360-backend/pyproject.toml
[tool.poetry]
name = "home360-backend"
version = "0.1.0"
description = "Backend service for Home360 mobile application"
authors = ["Your Name <your.email@example.com>"]

packages = [{ include = "app" }]

[tool.poetry.scripts]
start = "app.main:start"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
firebase-admin = "^6.2.0"
python-dotenv = "^1.0.0"
pydantic = "^2.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
black = "^23.9.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"